Another step towards the implementation of the Response class
This change adds the first bits of our response class. Right now it's
more like a container for info like the status code and the content
returned by the http call, but in the near future it will also perform
some operations on these values.

I had to remove the exceptions and coming up with another error
reporting system, because Qt doesn't play well with exceptions. This
stuff is not thread safe, maybe we'll work on it if we need to support
threads.

Also, a new method, called request was added to the Hollow class. It
takes the method and the url that the user wants to retrieve. Currently,
we're doing nothing but validating the url, using some QUrl stuff.

On the python side, we've got almost everything setup to call the
Hollow::request method and to convert its return to the Response python
class. Also, the SleepyHollow.get shortcut and some exceptions were
added to the binding.

This effort is docummented in the issue #4