Various improvements to avoid SleepyHollow.request to hang indefinitely
The WebPage::handleNetworkReplies method now compares the replied url to
the right url, We've added a safety check on the python C class to make
sure that it will raise an error even if the Error::last() method
doesn't return anything. The last change aiming to fix hangs was
overriding the method WebPage::shouldInterruptJavaScript() and make it
always return false.

Also, the QApplication is back to the Hollow class and it's created just
once.