Fixing the race condition that avoided us to exit the event loop
We've done a lot of stuff here. First, now we relay on two flags to say
that we should quit the main loop:

 1) We keep the track of all resources downloaded (we had to extend the
    QNetworkAccessManager to do that) and then, before saying that we're
    ready, we make sure that all the requested resources were downloaded
    (or at least responded).

 2) We're using the QWebFrame->loadFinished() signal again

So, when both of the above conditions are true, we say that the request
is done.

Another very important thing is that we're not using the app->exec(),
we're now using the app->processEvents(). Although, it breaks on Mac OS
X because of a known bug [0] that we need to get rid of.

[0] https://bugs.webkit.org/show_bug.cgi?id=29570