Refactoring the AT_WITH_QT macro to just find qmake
Now, instead of receiving the module list in this macro, we split it to
just find qmake and we introduced a new macro to actually load the right
modules in our .pro file.

We did that to make it possible to use the the QT_REQUIRED_QT_VERSION
before loading the modules. We needed that cause now we can compile
sleepy hollow against both qt4.x and 5.x, but when we compile against
5.x, we need to load a new module: webkitwidgets.