Not depending on the requested objs to determine if the page is loaded
We were tracking all the requested resources (by overriding the
QNetworkAccessManager.createRequest() method) and then checking if all
the requested objects were downloaded properly.

The problem with that approach is that if the resource is in the cache,
we won't have the chance to update our internal list, and thus making
this information unreliable.

This commit adds an option to disable the cache and exposes all the
downloaded objects, this way the user of this library can check which
resources were downloaded or not manually.