Implementing support to HTTP basic auth:
New Types:
=============

* Created the type `UsernamePasswordPair`, a std::pair that stores
solely username and password, it is analog to a python tuple like
`("john", "123")` becomes `std::make_pair("johndoe", "123")`

* Created `ConfigPair` as an alias to `std::pair<std::string, bool>`

WebPage:
========

The WebPage class now takes an optional `UsernamePasswordPair` before
the `Config` parameter in its constructor.

The contents of the `UsernamePasswordPair` are be translated into
private attributes for username and password.

It now listens to the signal authenticationRequired from the network
access manager, and gives 2 attempts of setting the HTTP basic auth
credentials, closing the network reply otherwise.

Hollow:
=======

The hollow class now takes an `UsernamePasswordPair` as an optional
parameter and repasses it to the WebPage class so it can retry the authentication.

Although the WebPage will not use those credentials unless they are
wrong and another attempt is required.

The reason for that is that the method `Hollow::request` will use the
credentials to formulate the URL