Changing the behavior of Hollow::request
now it always returns the lastResponse() regardless of whether
WebPage::hasErrors() return value.

The reason for that is that QWebPage returns boolean in
the loadFinished signal:
(http://qt-project.org/doc/qt-5.0/qtwebkit/qwebpage.html#loadFinished)

But it doesn't provide a way to get the error description and raise
the exception appropriately.

As a test, without this code if sleepy hollow is used to go to
http://yipit.com it would raise: `Error: both response and error are
null`

By returning the last response the user at least has the opportunity
to debug javascript errors, the generated html and so forth.